Move operations for silencing into an object
