+++ tests/linters/fixit_test.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 
 from silence_lint_error.linters.fixit import Fixit
-from silence_lint_error.silence_lint_error import Violation
+from silence_lint_error.silencing import Violation
 
 
 class TestFixit:
